requirements.txt
flask
python-dotenv
openai
requests
aiohttp
pydantic
langchain
tiktoken
langchain_openai
chromadb